+++ armstrong/esi/utils.py
